"Too many words in string" should be a warning, not an error
By being a warning and saying "will never match", it appears as if ChatScript will fail to match the given pattern, but this is not true. In our instance, we know exactly what is being sent in, and it should match exactly as given.

Therefore, this should be a warning for bot quality guidance and not an error that completely blocks the build.